Create setup.js for registering slash commands
First half of #30. Also, the log function is now a separate file so that both programs can use it.
